Move mount logic into root package
